requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.24.3
requests==2.31.0

# Blockchain & Web3
web3==6.11.3
eth-utils==2.3.0

# Environment & Configuration
python-dotenv==1.0.0

# Machine Learning (Lightweight)
scikit-learn==1.3.2

# Visualization (Minimal)
matplotlib==3.8.2
seaborn==0.13.0

# Development & Testing
jupyter==1.0.0
pytest==7.4.3

# Data Validation
cerberus==1.3.5

# Progress Tracking
tqdm==4.66.1

# Time Handling
python-dateutil==2.8.2
